requirements.txt
fastapi
uvicorn
scikit-learn
joblib
requests
psycopg2-binary